+++ src/components/Dropdown/index.ts
@@ -1,0 +1,2 @@
+export { default as Dropdown } from './Dropdown';
+export type { IDropdown } from './types';
